+++ GHEtool/VariableClasses/LoadData/__init__.py
@@ -1,3 +1,4 @@
 from .GeothermalLoad import *
 from .BuildingLoad import *
 from .Baseclasses import _LoadData, _LoadDataBuilding
+from .Cluster import Cluster